+++ lib/flipper_app.dart
@@ -1,0 +1,16 @@
+import 'package:flipper_login/login_view.dart';
+import 'package:flutter/material.dart';
+
+class FlipperApp extends StatelessWidget {
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: LoginView(),
+    );
+  }
+}
