+++ lib/main.dart
@@ -1,20 +1,7 @@
+import 'package:flipper/flipper_app.dart';
 import 'package:flutter/material.dart';
-import 'package:flipper_login/flipper_login.dart';
+import 'package:flipper_login/login_view.dart';
 
 void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: LoginView(),
-    );
-  }
+  runApp(FlipperApp());
 }
